requirements.txt
python-dotenv == 0.21.0
aiogram == 2.22.1
PyQiwiP2P == 2.0.6
openpyxl == 3.0.10

peewee == 3.15.3